feat: add per-page title metadata (fixes #165)
Each page previously rendered the same <title>Virtual Bookshelf</title>.
The root layout already has the title template; split login, signup, and
dashboard into a thin server-component page (which exports metadata) and
a companion Client component, so Next.js can apply the template to
produce "Sign in | Virtual Bookshelf", "Create account | Virtual Bookshelf",
and "My shelf | Virtual Bookshelf".
